Change return type of BmrCalculator interface to float

+++ src/BmrCalculator.php
@@ -4,5 +4,5 @@ namespace App;
 
 interface BmrCalculator
 {
-	public function calculateBmr(int $weight, int $height, int $age, int $gender): int;
+	public function calculateBmr(int $weight, int $height, int $age, int $gender): float;
 }
